Implement union verifier
